# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yappy Region Sync",
    "description": "Diff de DB (tablas/SP) y de parámetros SSM/Secrets entre regiones de origen y destino",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/params/diff": {
      "post": {
        "tags": [
          "params"
        ],
        "summary": "Api Params Diff",
        "operationId": "params_diff",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParamsDiffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParamsDiffResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/params/apply": {
      "post": {
        "tags": [
          "params"
        ],
        "summary": "Api Params Apply",
        "operationId": "params_apply",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyParamsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParamsApplyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/params/apply-execute": {
      "post": {
        "tags": [
          "params"
        ],
        "summary": "Api Params Apply Execute",
        "operationId": "params_apply_execute",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteParamsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteParamsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/params/multi": {
      "post": {
        "tags": [
          "params"
        ],
        "summary": "Api Params Multi",
        "operationId": "params_multi",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMultiParamsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParamsMultiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/params/get": {
      "get": {
        "tags": [
          "params"
        ],
        "summary": "Api Params Get",
        "operationId": "params_get",
        "parameters": [
          {
            "name": "env",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Env"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterReadInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/params/read": {
      "post": {
        "tags": [
          "params"
        ],
        "summary": "Api Params Read",
        "description": "Read the same keys across every selected environment.\n\nOne unreachable environment must not hide the values of the others, so each\nenvironment is attempted independently and a failure becomes error rows\ncarrying that environment's name.",
        "operationId": "params_read",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_params_read"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParamsReadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplyParamsRequest": {
        "properties": {
          "env_a": {
            "type": "string",
            "title": "Env A"
          },
          "env_b": {
            "type": "string",
            "title": "Env B"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "new_value": {
            "type": "string",
            "title": "New Value"
          },
          "value_type": {
            "type": "string",
            "title": "Value Type",
            "default": "String"
          },
          "with_secret": {
            "type": "boolean",
            "title": "With Secret",
            "default": false
          },
          "new_secret_value": {
            "type": "string",
            "title": "New Secret Value",
            "default": ""
          },
          "write_secret": {
            "type": "boolean",
            "title": "Write Secret",
            "default": false
          },
          "write_param": {
            "type": "boolean",
            "title": "Write Param",
            "default": true
          },
          "target": {
            "type": "string",
            "title": "Target",
            "default": "a"
          }
        },
        "type": "object",
        "required": [
          "env_a",
          "env_b",
          "service",
          "name",
          "new_value"
        ],
        "title": "ApplyParamsRequest"
      },
...
      "ExecuteParamsRequest": {
        "properties": {
          "env_a": {
            "type": "string",
            "title": "Env A"
          },
          "env_b": {
            "type": "string",
            "title": "Env B"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "op": {
            "type": "string",
            "title": "Op",
            "default": "update"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "new_value": {
            "type": "string",
            "title": "New Value",
            "default": ""
          },
          "value_type": {
            "type": "string",
            "title": "Value Type",
            "default": "String"
          },
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": false
          },
          "with_secret": {
            "type": "boolean",
            "title": "With Secret",
            "default": false
          },
          "new_secret_value": {
            "type": "string",
            "title": "New Secret Value",
            "default": ""
          },
          "write_secret": {
            "type": "boolean",
            "title": "Write Secret",
            "default": false
          },
          "write_param": {
            "type": "boolean",
            "title": "Write Param",
            "default": true
          },
          "target": {
            "type": "string",
            "title": "Target",
            "default": "a"
          }
        },
        "type": "object",
        "required": [
          "env_a",
          "env_b",
          "service",
          "name"
        ],
        "title": "ExecuteParamsRequest"
      },
      "ExecuteParamsResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "env_a": {
            "type": "string",
            "title": "Env A"
          },
          "env_b": {
            "type": "string",
            "title": "Env B"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "op": {
            "type": "string",
            "title": "Op"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepResultInfo"
            },
            "type": "array",
            "title": "Steps",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "ok",
          "message",
          "env_a",
          "env_b",
          "service",
          "op",
          "name"
        ],
        "title": "ExecuteParamsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParameterReadInfo": {
        "properties": {
          "env": {
            "type": "string",
            "title": "Env"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "value_type": {
            "type": "string",
            "title": "Value Type"
          }
        },
        "type": "object",
        "required": [
          "env",
          "name",
          "value",
          "value_type"
        ],
        "title": "ParameterReadInfo"
      },
      "ParamsApplyResponse": {
        "properties": {
          "env_a": {
            "type": "string",
            "title": "Env A"
          },
          "env_b": {
            "type": "string",
            "title": "Env B"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "script": {
            "type": "string",
            "title": "Script"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepCommandInfo"
            },
            "type": "array",
            "title": "Steps",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "env_a",
          "env_b",
          "service",
          "name",
          "script"
        ],
        "title": "ParamsApplyResponse"
      },
      "ParamsDiffRequest": {
        "properties": {
          "env_a": {
            "type": "string",
            "title": "Env A"
          },
          "env_b": {
            "type": "string",
            "title": "Env B"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "include_deletes": {
            "type": "boolean",
            "title": "Include Deletes",
            "default": false
          },
          "with_secret": {
            "type": "boolean",
            "title": "With Secret",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "env_a",
          "env_b",
          "service",
          "name"
        ],
        "title": "ParamsDiffRequest"
      },
      "ParamsDiffResponse": {
        "properties": {
          "env_a": {
            "type": "string",
            "title": "Env A"
          },
          "env_b": {
            "type": "string",
            "title": "Env B"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "value_a": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value A"
          },
          "value_b": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value B"
          },
          "value_type_a": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Type A"
          },
          "value_type_b": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Type B"
          },
          "is_json": {
            "type": "boolean",
            "title": "Is Json",
            "default": false
          },
          "changes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Changes",
            "default": []
          },
          "patch_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patch Value"
          },
          "script": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Script"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes",
            "default": []
          },
          "pair": {
            "type": "boolean",
            "title": "Pair",
            "default": false
          },
          "secret_value_a": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Value A"
          },
          "secret_value_b": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Value B"
          },
          "secret_changes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Secret Changes",
            "default": []
          },
          "secret_patch_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Patch Value"
          },
          "param_apply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param Apply"
          },
          "secret_apply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Apply"
          },
          "param_status": {
            "type": "string",
            "title": "Param Status",
            "default": ""
          },
          "secret_status": {
            "type": "string",
            "title": "Secret Status",
            "default": ""
          },
          "param_needs_write": {
            "type": "boolean",
            "title": "Param Needs Write",
            "default": false
          },
          "secret_needs_write": {
            "type": "boolean",
            "title": "Secret Needs Write",
            "default": false
          },
          "steps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Steps",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "env_a",
          "env_b",
          "service",
          "name",
          "status"
        ],
        "title": "ParamsDiffResponse"
      },
...
      "StepCommandInfo": {
        "properties": {
          "step": {
            "type": "string",
            "title": "Step"
          },
          "command": {
            "type": "string",
            "title": "Command"
          }
        },
        "type": "object",
        "required": [
          "step",
          "command"
        ],
        "title": "StepCommandInfo"
      },
...
    }
  }
}